Move all test files to /test following convention
