+++ Data-Structures/Vectors/test/Vector2.test.js
@@ -1,4 +1,4 @@
-import { Vector2 } from './Vector2.js'
+import { Vector2 } from '../Vector2.js'
 
 describe('Vector2', () => {
   describe('#equalsExactly', () => {
